requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.61.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pefile==2024.8.26
pillow==12.1.0
playwright>=1.40.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdfium2==5.3.0
PyPDF2>=3.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx>=1.1.0
python-dotenv>=1.0.0
pytz==2025.2
pywin32-ctypes==0.2.3
RapidFuzz==3.14.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
schedule>=1.2.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition>=3.10.0
pydub>=0.25.1
docx2pdf>=0.1.8
tenacity==9.1.2
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
